Add files via upload
Summary: Summary: 
Overall, the data set presents a robust set of values--most columns included complete values. Although 'Population' presented the most missing values (about 22 percent), and we used the mean value to fill in the missing values, we see that the correlation of 'Population' and 'Life expectancy' is neutral. That being said, we observe four big picture data points. First, we see in Figure 1 that the 'Status' of a country--whether its 'Developing' or 'Developed'--the percentage of expenditure is higher in the latter group.  Second, we see in Figure 2 that 'Life expectancy' increases with the percent of government expenditure on its people's welfare. There's a positive correlation between the two variables of .38.  Third, we see that the diseases ('Alcohol', 'Hepatitis B', 'Measles ','Polio', 'Diphtheria'  and 'HIV/AIDS') show a moderate correlation with 'Life expectancy', or the number of years one lives in a given country. Interestingly enough, 'HIV/AIDS shows the highest correlation of .59 with 'Life expectancy', which is the most modern disease to plague society. I'm curious to see the correlation of Novel COVID-19 in upcoming WHO data. Fourth, we see in Figure 4 (Life Expectancy Per Country) that a 'Developing' country like, Afghanistan or Coite d'oivre, has a lower life expectancy range than the 'Developed' country, like Kuwait or Qatar, which range between 73.2-74.7 and 76.3-78.6 respectively.. Next time, I would like to break the countries into groups by 'Status' to show the state of development's interaction with 'Life expectancy', or the number of years one lives in a given country.